+++ src/Bootstrap/LoadConfiguration.php
@@ -1,0 +1,60 @@
+<?php namespace Orchestra\Testbench\Bootstrap;
+
+use Illuminate\Config\Repository;
+use Symfony\Component\Finder\Finder;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\Config\Repository as RepositoryContract;
+
+class LoadConfiguration
+{
+
+    /**
+     * Bootstrap the given application.
+     *
+     * @param  \Illuminate\Contracts\Foundation\Application  $app
+     * @return void
+     */
+    public function bootstrap(Application $app)
+    {
+        $items = [];
+
+        $app->instance('config', $config = new Repository($items));
+
+        $this->loadConfigurationFiles($app, $config);
+
+        date_default_timezone_set($config['app.timezone']);
+
+        mb_internal_encoding('UTF-8');
+    }
+
+    /**
+     * Load the configuration items from all of the files.
+     *
+     * @param  \Illuminate\Contracts\Foundation\Application  $app
+     * @param  \Illuminate\Contracts\Config\Repository  $config
+     * @return void
+     */
+    protected function loadConfigurationFiles(Application $app, RepositoryContract $config)
+    {
+        foreach ($this->getConfigurationFiles($app) as $key => $path) {
+            $config->set($key, require $path);
+        }
+    }
+
+    /**
+     * Get all of the configuration files for the application.
+     *
+     * @param  \Illuminate\Contracts\Foundation\Application  $app
+     * @return array
+     */
+    protected function getConfigurationFiles(Application $app)
+    {
+        $files = [];
+
+        foreach (Finder::create()->files()->name('*.php')->in($app->configPath()) as $file) {
+            $files[basename($file->getRealPath(), '.php')] = $file->getRealPath();
+        }
+
+        return $files;
+    }
+}