+++ docs/changes.md
@@ -8,6 +8,7 @@ title: Testbench Change Log
 ### v3.0.2@dev {#v3-0-2}
 
 * Timezone should be more explicit, and shouldn't attempt to set `date_default_timezone_set()` when timezone is `NULL`.
+* Rebind `Illuminate\Foundation\Bootstrap\LoadConfiguration` with `Orchestra\Testbench\Bootstrap\LoadConfiguration`.
 
 ### v3.0.1 {#v3-0-1}